assignment/task2/docker/requirements.txt
apache-airflow-providers-slack
apache-airflow-providers-apache-spark